Test fixes and code restructures.
